sensor/requirements.txt
requests
crccheck
pyserial